pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multiplayer-session-recorder"
description = "Multiplayer Session Recorder"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.8"
authors = [
    { name = "Multiplayer Authors", email = "hello@multiplayer.app" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: OpenTelemetry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "opentelemetry-exporter-otlp-proto-http >= 1.29.0",
    "opentelemetry-exporter-otlp-proto-grpc >= 1.29.0",
    "opentelemetry-api >= 1.29.0",
    "opentelemetry-sdk >= 1.29.0",
    "aiohttp >= 3.8.0"
]
dynamic = [
    "version",
]

[project.optional-dependencies]
flask = [
  "flask>=2.0",
  "opentelemetry-instrumentation-flask>=0.43b0"
]
django = [
  "django>=3.2",
  "opentelemetry-instrumentation-django>=0.43b0"
]
all = [
  "flask>=2.0",
  "django>=3.2",
  "opentelemetry-instrumentation-flask>=0.43b0",
  "opentelemetry-instrumentation-django>=0.43b0"
]

[project.urls]
Homepage = "https://github.com/multiplayer-app/multiplayer-session-recorder-python"
Repository = "https://github.com/multiplayer-app/multiplayer-session-recorder-python"

[tool.hatch.version]
path = "src/version/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = ["src", "README.md", "LICENSE"]

[tool.hatch.build.targets.wheel]
packages = ["src/multiplayer_session_recorder"]
